show regulatory metadata on click
