+++ frontend/src/containers/Map.js
@@ -13,8 +13,8 @@ import VesselCardOverlay from '../components/overlays/VesselCardOverlay'
 import VesselTrackCardOverlay from '../components/overlays/VesselTrackCardOverlay'
 import TrackTypeCardOverlay from '../components/overlays/TrackTypeCardOverlay'
 import MapVesselAnimation from '../components/map/MapVesselAnimation'
-import ShowRegulatoryMetadata from '../components/map/ShowRegulatoryMetadata'
-const hitPixelTolerance = 3
+import { HIT_PIXEL_TO_TOLERANCE } from '../constants/constants'
+console.log('HIT_PIXEL_TO_TOLERANCE ' + HIT_PIXEL_TO_TOLERANCE)
 
 const Map = () => {
   const gears = useSelector(state => state.gear.gears)
@@ -24,7 +24,6 @@ const Map = () => {
   const [historyMoveTrigger, setHistoryMoveTrigger] = useState({})
   const [currentFeature, setCurrentFeature] = useState(null)
   const [mapMovingAndZoomEvent, setMapMovingAndZoomEvent] = useState(null)
-  const [mapClickEvent, setMapClickEvent] = useState(null)
   const [handlePointerMoveEventPixel, setHandlePointerMoveEventPixel] = useState(null)
 
   const handleMovingAndZoom = () => {
@@ -36,17 +35,10 @@ const Map = () => {
     setMapMovingAndZoomEvent({ dummyUpdate: true })
   }
 
-  const handleMapClick = (event, map) => {
-    if (event && map) {
-      const feature = map.forEachFeatureAtPixel(event.pixel, feature => feature, { hitTolerance: hitPixelTolerance })
-      setMapClickEvent({ feature })
-    }
-  }
-
   const handlePointerMove = (event, map) => {
     if (event && map) {
       const pixel = map.getEventPixel(event.originalEvent)
-      const feature = map.forEachFeatureAtPixel(pixel, feature => feature, { hitTolerance: hitPixelTolerance })
+      const feature = map.forEachFeatureAtPixel(pixel, feature => feature, { hitTolerance: HIT_PIXEL_TO_TOLERANCE })
       setHandlePointerMoveEventPixel(event.pixel)
       if (feature && feature.getId()) {
         setCurrentFeature(feature)
@@ -62,12 +54,11 @@ const Map = () => {
         <BaseMap
           handleMovingAndZoom={handleMovingAndZoom}
           handlePointerMove={handlePointerMove}
-          handleMapClick={handleMapClick}
+          showCoordinates={true}
         >
             <MapVesselAnimation
               mapMovingAndZoomEvent={mapMovingAndZoomEvent}
-              mapClickEvent={mapClickEvent} />
-            <ShowRegulatoryMetadata mapClickEvent={mapClickEvent} />
+            />
             <MapHistory
               shouldUpdateView={shouldUpdateView}
               setShouldUpdateView={setShouldUpdateView}
